requirements.txt
crewai[tools]==0.65.2
langchain==0.2.16
langchain-ollama==0.1.3
python-dotenv==1.0.1
googlesearch-python==1.2.5
unstructured==0.15.13
pylint==3.3.1
black==24.8.0
